requirements.txt
flask
pandas
python-jobspy
pyyaml
